Commiting a lot of changes. Not sure what they all were. Stuff after first 12 person runthrough.

+++ app/controllers/user_sessions_controller.rb
@@ -14,7 +14,7 @@ class UserSessionsController < InheritedResources::Base
   end
 
   def destroy
-    current_user_session.destroy
+    current_user_session.destroy unless current_user_session.nil?
     redirect_to root_path
   end
 end
